feat: API 변경

+++ src/main/java/com/example/publicdataserver/domain/restaurant/Restaurant.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.ColumnDefault;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -32,7 +33,7 @@ public class Restaurant {
     private String x;
     private String y;
 
-    private Integer rating;
+    private Double rating;
     private String storeId;
 
     @Column(columnDefinition = "TEXT")
